+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -98,6 +98,10 @@ class AppTableTest extends TableTestCase
             $this->assertEquals('2017-03-14 20:19:00', $entity->get('created')->i18nFormat('yyyy-MM-dd HH:mm:ss'));
             $Table->delete($entity);
         }
+
+        $entity = $Table->save($Table->newEntity(['created' => null] + $example));
+        $this->assertInstanceOf(Time::class, $entity->get('created'));
+        $Table->delete($entity);
     }
 
     /**