+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## 2.29 branch
 ### 2.29.4
 * `SitemapBuilder::getMethods()` returns a collection and it's now public;
+* `StaticPage::all()` returns a collection;
 * fixed `TestSuite` classes, `MenuBuilderHelper` and some template files, improved
     to use with other plugins;
 * fixed old bug on loading theme plugins;